Created two folders for presentation tomorrow
Created two folders for tomorrows presentation. SmallDemo is where I will show that it works with the merge sort solution with small inputs and not for large, and the LargeDemo will be for showing that the distance values are going to the be same for both outputs (distance and knn solutions).